test(frontend): Added back 2 previously steps, to the right place this time... Added wait time for ready button
